CDD-1554: Made changes based on PR review comments
