+++ test/component/CustomsDeclarationClearanceSpec.scala
@@ -105,8 +105,6 @@ class CustomsDeclarationClearanceSpec extends ComponentTestSpec with AuditServic
 
   }
 
-  //*** UNHAPPY PATH SPECS ***
-
   feature("The API handles errors as expected") {
 
     scenario("Response status 400 when user submits an xml payload that does not adhere to schema having multiple errors") {
