+++ test/component/CustomsDeclarationSubmissionSpec.scala
@@ -154,8 +154,6 @@ class CustomsDeclarationSubmissionSpec extends ComponentTestSpec with AuditServi
 
   }
 
-  //*** UNHAPPY PATH SPECS ***
-
   feature("The API handles errors as expected") {
 
     scenario("Response status 400 when user submits an xml payload that does not adhere to schema having multiple errors") {
